BBDD/requirements.txt
mysql-connector==2.1.6
pyodbc==4.0.24
